cpymo-tool strip (#47)
* asset_analyzer

* Update CI

* Update ci.yml

* Update ci.yml

* Update CMakeLists.txt

* Try fix wii build

* Try fix shit

* Try fix wii build

* Try fix wii

* Try fix wii build

* Try fix wii build

* Try fix wii build

* Try fix wii build

* Try fix uwp build

* update uwp cerf

* Update ci.yml

* Update CPyMO.vcxproj

---------


* analyzer

* Bug fixedd.

* Bug fixed.

* cpymo_tool_asset_filter.h

* asset_filter design

* Add someshit.

* add cpymo_tool_package_packer api

* rewrite cpymo pack in cpymo tool pack api

* step by step.

* stricter pack api

* filter api impl basiclly

* processing

* cpymo_tool_asset_filter_run_single

* filter api finished

* cpymo-tool strip

* remove pymo-strip.ps1

* Bug fixed

* cpymo_tool_gameconfig

* bug fixed

* bug fixed.

* ffmpeg-searcher

* add some shit functions

* convert features.

* step by step(

* step by step

* convert finished, wait for debug

* bug fixed.

* convert

* remove powershell version convert

* bug fixed

---------

+++ cpymo-tool/cpymo_tool_package.c
@@ -80,7 +80,7 @@ error_t cpymo_tool_package_packer_open(
 {
 	packer->current_file_count = 0;
 	packer->data_section_start_offset =
-		sizeof(uint32_t) + max_files_count * sizeof(cpymo_package_index);
+		(long)(sizeof(uint32_t) + max_files_count * sizeof(cpymo_package_index));
 	packer->index_section_start_offset = sizeof(uint32_t);
 	packer->max_file_count = max_files_count;
